Switch from Okio to kotlinx.io
There is a performance hit because of this change but it might enable
Kotlin/Native development

+++ pgkotlin/src/commonMain/kotlin/com/bnorm/pgkotlin/internal/msg/StatementType.kt
@@ -1,6 +1,6 @@
 package com.bnorm.pgkotlin.internal.msg
 
-enum class StatementType(val code: Int) {
-  Prepared('S'.toInt()),
-  Portal('P'.toInt()),
+enum class StatementType(val code: Byte) {
+  Prepared('S'.toByte()),
+  Portal('P'.toByte()),
 }
